unionfs: Open: return a seekableFile if file supports io.Seeker
This was discovered on Windows when using http.FileServer/serveContent:
Since on Windows MIME types are extracted from the registry, behavior
of serveContent may be different there as compared to Linux.
If the type of a file like with an extension like ".woff2"
cannot be read from the MIME-info database, serveContent reads
a chunk from the file to test for binary / utf-8 content, then it
tries to rewind the file making use of the file's io.Seeker.
If there is none, it fails. This change ensures that the Seeker is
made available, if present.